Added other entity classes. Fix: interaction with non-existent data

+++ models/lectureModel.go
@@ -1,14 +1,14 @@
 package models
 
 type Lecture struct {
-	Id        int    `json:"id"`
+	Id        int64  `json:"id"`
 	Date      string `json:"date"`
-	SubjectId int    `json:"subject_id"`
-	TeacherId int    `json:"teacher_id"`
+	SubjectId int64  `json:"subject_id"`
+	TeacherId int64  `json:"teacher_id"`
 }
 
 type LectureData struct {
 	Date      string `json:"date"`
-	SubjectId int    `json:"subject_id"`
-	TeacherId int    `json:"teacher_id"`
+	SubjectId int64  `json:"subject_id"`
+	TeacherId int64  `json:"teacher_id"`
 }
